requirements.txt
elasticsearch==5.4.0
openai==1.30.3
streamlit==1.36.0
python-dotenv==1.0.1